convert all positions to Vec (float vectors)

+++ nova2d/nova-zero/src/core/BoundUser.h
@@ -54,8 +54,8 @@ namespace novazero
 				bool outOfBounds = false;
 				if (sprite)
 				{
-					int x = sprite->GetX();
-					int y = sprite->GetY();
+					float x = sprite->GetX();
+					float y = sprite->GetY();
 					int w = sprite->GetWidth();
 					int h = sprite->GetHeight();
